start

+++ workflow/scripts/fire-null-distribution.py
@@ -87,7 +87,7 @@ def fire_scores_per_chrom(
     min_allowed_q=0.01,
     min_coverage=4,
 ):
-    fire_scores = np.zeros(chrom_length, dtype=np.float64)
+    fire_scores = np.zeros(np.float64(chrom_length), dtype=np.float64)
 
     multi = -50.0  # a multi of -50 and a min_allowed_q of 0.01 gives a max score of 100
     max_add = multi * np.log10(min_allowed_q)
